chore: rename defaultPlugin
