+++ packages/create-two-scripts/utils/paths.ts
